+++ packages/browser-extension/src/modules/service/utils/get-summary-message.ts
@@ -1,6 +1,7 @@
-import { SyncResponse } from "../handlers/handle-sync";
+import { SyncContentResponse } from "../handlers/handle-sync-content";
+import { SyncMetadataResponse } from "../handlers/handle-sync-metadata";
 
-export function getSummaryMessage(syncResponse: SyncResponse): string {
+export function getSummaryMessage(syncResponse: SyncContentResponse): string {
   let summaryMessage = "";
   if (syncResponse.addedIds.length) summaryMessage += ` ${syncResponse.addedIds.length} added`;
   if (syncResponse.updatedIds.length) summaryMessage += ` ${syncResponse.updatedIds.length} updated`;
@@ -9,4 +10,17 @@ export function getSummaryMessage(syncResponse: SyncResponse): string {
   if (!summaryMessage.length) summaryMessage += " No change";
 
   return `Sync items... Success! (${summaryMessage.trim()})`;
+}
+
+export function getShortSummaryMessage(contentSummary: SyncContentResponse, metadataSummary?: SyncMetadataResponse): string {
+  const totalContentChange = contentSummary.addedIds.length + contentSummary.deletedIds.length + contentSummary.updatedIds.length;
+  const totalMetadataChange = metadataSummary?.itemTypeCount;
+
+  let summaryMessage = "";
+
+  if (totalContentChange > 0) summaryMessage += `${totalContentChange} content changes`;
+  if (totalMetadataChange && totalContentChange > 0) summaryMessage += ` ${totalMetadataChange} metadata changes`;
+  if (!summaryMessage.length) summaryMessage = "No change";
+
+  return `Sync... Success! (${summaryMessage.trim()})`;
 }